+++ packages/components/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.10.0](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.9.0...@ttoss/components@2.10.0) (2025-12-01)
+
+### Features
+
+- adapt component to use badge instead of subtitle ([#788](https://github.com/ttoss/ttoss/issues/788)) ([0c3c7d8](https://github.com/ttoss/ttoss/commit/0c3c7d850f2b9fd817afa7b0e0ed4c5e7fa6a9c5))
+
+# [2.9.0](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.8.0...@ttoss/components@2.9.0) (2025-11-28)
+
+### Features
+
+- **SpotlightCard:** update variant options and enhance tests for primary variant ([#784](https://github.com/ttoss/ttoss/issues/784)) ([3632fb2](https://github.com/ttoss/ttoss/commit/3632fb21fa2176ba3e4d02c63606ffab2e52a762))
+
+# [2.8.0](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.7.11...@ttoss/components@2.8.0) (2025-11-27)
+
+### Features
+
+- add SpotlightCard component with responsive design and animations ([#776](https://github.com/ttoss/ttoss/issues/776)) ([47d02fd](https://github.com/ttoss/ttoss/commit/47d02fd22f97ab51d6e70f3a24356da047fe45c0))
+
 ## [2.7.11](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.7.10...@ttoss/components@2.7.11) (2025-11-26)
 
 **Note:** Version bump only for package @ttoss/components