+++ packages/react-auth/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.72](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.71...@ttoss/react-auth@2.6.72) (2025-12-01)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
+## [2.6.71](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.70...@ttoss/react-auth@2.6.71) (2025-12-01)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
+## [2.6.70](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.69...@ttoss/react-auth@2.6.70) (2025-11-28)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
+## [2.6.69](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.68...@ttoss/react-auth@2.6.69) (2025-11-28)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
+## [2.6.68](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.67...@ttoss/react-auth@2.6.68) (2025-11-27)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
 ## [2.6.67](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@2.6.66...@ttoss/react-auth@2.6.67) (2025-11-26)
 
 **Note:** Version bump only for package @ttoss/react-auth